move generated to another folder
