+++ blockchain/beacon/build.rs
@@ -1,13 +1,10 @@
 // Copyright 2020 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
-
-// Copyright 2020 ChainSafe Systems
-// SPDX-License-Identifier: Apache-2.0, MIT
 use protoc_rust_grpc::Codegen;
 fn main() {
     Codegen::new()
         .includes(&["proto", "proto/api-common-protos"])
-        .out_dir("src")
+        .out_dir("src/drand_api")
         .inputs(&["proto/api.proto", "proto/common.proto"])
         .rust_protobuf(true)
         .run()
