+++ blockchain/beacon/src/drand_api/common.rs
