+++ blockchain/beacon/src/drand_api/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod api;
+pub(crate) mod api_grpc;
+pub(crate) mod common;
