+++ blockchain/beacon/src/lib.rs
@@ -1,13 +1,11 @@
 // Copyright 2020 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-mod api;
-mod api_grpc;
 mod beacon_entries;
-mod common;
 mod drand;
 mod group;
 mod identity;
+mod drand_api;
 
 pub use beacon_entries::*;
 pub use drand::*;
